requirements.txt
urllib3
networkx
osmnx
staticmap
haversine
requests
python-telegram-bot
